requirements.txt
numpy
Pillow
scikit-learn
opencv-python
qrcode
supabase
fastapi
uvicorn
python-dotenv
pydantic